alternative jumble sort
The given jumble sort solution seems problematic with respect to efficiency: bubble sort is O(n^2) and each call to Array#index is O(n). This commit adds an alternative solution. (Granted, the use of sort_by is bad if our goal is to test the candidate's computer-science fundamentals, but is fine if instead we're interested in the candidate's knowledge of Ruby Array methods.)